# feat: speed up `PrefixRecord` loading