--- a/Cargo.toml
+++ b/Cargo.toml
@@ -107,6 +107,7 @@ plist = "1"
 purl = { version = "0.1.3", features = ["serde"] }
 quote = "1.0.37"
 rand = "0.8.5"
+rayon = "1.10.0"
 reflink-copy = "0.1.20"
 regex = "1.11.1"
 reqwest = { version = "0.12.9", default-features = false }