--- a/Cargo.toml
+++ b/Cargo.toml
@@ -33,27 +33,27 @@ async-compression = { version = "0.4.17", features = [
     "zstd",
 ] }
 async-fd-lock = "0.2.0"
-fs4 = "0.11.0"
+fs4 = "0.12.0"
 async-trait = "0.1.83"
 axum = { version = "0.7.7", default-features = false, features = [
     "tokio",
     "http1",
 ] }
 base64 = "0.22.1"
-bindgen = "0.69.5"
+bindgen = "0.71.1"
 blake2 = "0.10.6"
 bytes = "1.8.0"
-bzip2 = "0.4.4"
+bzip2 = "0.5.0"
 cache_control = "0.2.0"
 cfg-if = "1.0"
-chrono = { version = "0.4.38", default-features = false, features = [
+chrono = { version = "0.4.39", default-features = false, features = [
     "std",
     "serde",
     "alloc",
 ] }
-clap = { version = "4.5.20", features = ["derive"] }
+clap = { version = "4.5.23", features = ["derive"] }
 cmake = "0.1.51"
-console = { version = "0.15.8", features = ["windows-console-colors"] }
+console = { version = "0.15.10", features = ["windows-console-colors"] }
 criterion = "0.5"
 dashmap = "6.1.0"
 difference = "2.0.0"
@@ -74,11 +74,11 @@ google-cloud-auth = { version = "0.17.1", default-features = false }
 google-cloud-token = "0.1.2"
 hex = "0.4.3"
 hex-literal = "0.4.1"
-http = "1.1"
+http = "1.2"
 http-cache-semantics = "2.1.0"
 humansize = "2.1.3"
 humantime = "2.1.0"
-indexmap = "2.6.0"
+indexmap = "2.7.0"
 indicatif = "0.17.8"
 insta = { version = "1.41.1" }
 itertools = "0.13.0"
@@ -99,7 +99,7 @@ once_cell = "1.20.2"
 ouroboros = "0.18.4"
 parking_lot = "0.12.3"
 pathdiff = "0.2.2"
-pep440_rs = { version = "0.7.2" }
+pep440_rs = { version = "0.7.3" }
 pep508_rs = { version = "0.9.1" }
 percent-encoding = "2.3.1"
 pin-project-lite = "0.2.15"
@@ -119,7 +119,7 @@ rmp-serde = { version = "1.3.0" }
 rstest = { version = "0.23.0" }
 rstest_reuse = "0.7.0"
 simd-json = { version = "0.14.2", features = ["serde_impl"] }
-serde = { version = "1.0.214" }
+serde = { version = "1.0.216" }
 serde_json = { version = "1.0.132" }
 serde_repr = "0.1"
 serde-value = "0.7.0"
@@ -138,23 +138,23 @@ smallvec = { version = "1.13.2", features = [
 strum = { version = "0.26.3", features = ["derive"] }
 superslice = "1.0.0"
 syn = "2.0.86"
-sysinfo = "0.32.0"
+sysinfo = "0.33.0"
 tar = "0.4.42"
 tempdir = "0.3.7"
 tempfile = "3.13.0"
 temp-env = "0.3.6"
 test-log = "0.2.16"
-thiserror = "1.0"
-tokio = { version = "1.41.0", default-features = false }
-tokio-stream = "0.1.16"
-tokio-util = "0.7.12"
-tower = { version = "0.5.1", default-features = false }
+thiserror = "2.0"
+tokio = { version = "1.42.0", default-features = false }
+tokio-stream = "0.1.17"
+tokio-util = "0.7.13"
+tower = { version = "0.5.2", default-features = false }
 tower-http = { version = "0.6.1", default-features = false }
 tracing = "0.1.40"
 tracing-subscriber = { version = "0.3.18", default-features = false }
 tracing-test = { version = "0.2.5" }
 trybuild = { version = "1.0.101" }
-typed-path = { version = "0.9.3" }
+typed-path = { version = "0.10.0" }
 url = { version = "2.5.2" }
 uuid = { version = "1.11.0", default-features = false }
 walkdir = "2.5.0"