--- a/Cargo.toml
+++ b/Cargo.toml
@@ -23,7 +23,7 @@ tag-prefix = ""
 lto = true
 
 [workspace.dependencies]
-anyhow = "1.0.92"
+anyhow = "1.0.94"
 archspec = "0.1.3"
 assert_matches = "1.5.0"
 async-compression = { version = "0.4.17", features = [