--- a/crates/rattler_conda_types/Cargo.toml
+++ b/crates/rattler_conda_types/Cargo.toml
@@ -10,6 +10,9 @@ repository.workspace = true
 license.workspace = true
 readme.workspace = true
 
+[features]
+default = ["rayon"]
+
 [dependencies]
 chrono = { workspace = true }
 file_url = { path = "../file_url", version = "0.2.0" }