--- a/crates/rattler_conda_types/benches/prefix_record_from_path.rs
+++ b/crates/rattler_conda_types/benches/prefix_record_from_path.rs
@@ -1,5 +1,6 @@
+use rattler_conda_types::RecordFromPath;
 use std::{
-    fs,
+    fs::{self},
     path::{Path, PathBuf},
 };
 
@@ -27,6 +28,18 @@ fn load_as_package_record(dir: &Path) -> Vec<PackageRecord> {
 
 fn criterion_benchmark(c: &mut Criterion) {
     let test_dir = PathBuf::from(env!("CARGO_MANIFEST_DIR")).join("../../test-data/conda-meta");
+
+    let mut super_long_file: PrefixRecord =
+        PrefixRecord::from_path(test_dir.join("tk-8.6.13-h5083fa2_1.json")).unwrap();
+    // duplicate data until we have 20k paths
+    let files = super_long_file.files.clone();
+    while super_long_file.files.len() < 20_000 {
+        super_long_file.files.extend(files.clone());
+    }
+
+    let tempfile = tempfile::NamedTempFile::new().unwrap();
+    serde_json::to_writer(&tempfile, &super_long_file).unwrap();
+
     c.bench_function("load_prefix_record_serially", |b| {
         b.iter(|| {
             process_json_files_from_dir(&test_dir);
@@ -37,6 +50,17 @@ fn criterion_benchmark(c: &mut Criterion) {
     });
     c.bench_function("load_as_package_record", |b| {
         b.iter(|| load_as_package_record(&test_dir));
+    });
+
+    let path = tempfile.path();
+    c.bench_function("load_long_prefix_record", |b| {
+        b.iter(|| black_box(PrefixRecord::from_path(path).unwrap()));
+    });
+
+    c.bench_function("load_long_package_record", |b| {
+        b.iter(|| {
+            black_box(PackageRecord::from_path(path).unwrap());
+        });
     });
 }