--- a/crates/rattler_conda_types/src/lib.rs
+++ b/crates/rattler_conda_types/src/lib.rs
@@ -50,7 +50,8 @@ pub use repo_data::{
     compute_package_url,
     patches::{PackageRecordPatch, PatchInstructions, RepoDataPatch},
     sharded::{Shard, ShardedRepodata, ShardedSubdirInfo},
-    ChannelInfo, ConvertSubdirError, PackageRecord, RepoData, ValidatePackageRecordsError,
+    ChannelInfo, ConvertSubdirError, PackageRecord, RecordFromPath, RepoData,
+    ValidatePackageRecordsError,
 };
 pub use repo_data_record::RepoDataRecord;
 pub use run_export::RunExportKind;