--- a/crates/tools/src/libsolv_bindings.rs
+++ b/crates/tools/src/libsolv_bindings.rs
@@ -105,6 +105,7 @@ pub fn generate(mode: Mode) -> anyhow::Result<()> {
 
     // Define the contents of the bindings and how they are generated
     let bindings = bindgen::Builder::default()
+        .rust_target("1.81.0".parse().unwrap())
         .clang_arg(format!("-I{}", temp_include_dir.path().display()))
         .ctypes_prefix("libc")
         .header(libsolv_path.join("src/solver.h").to_str().unwrap())