I have tried out rayon to parallelize the loading of `PrefixRecord` files. The results look pretty good:

```
load_serially           time:   [10.369 ms 10.403 ms 10.440 ms]
load_as_prefix_record   time:   [697.90 ns 699.63 ns 701.53 ns]
load_as_package_record  time:   [698.94 ns 700.45 ns 702.11 ns]
```

Unfortunately loading the subset (`PackageRecord`) doesn't seem to be a big improvement.