@wolfv 2024-12-16
For very long PrefixRecord's it does seem to make a difference (20k files):

```
load_long_prefix_record time:   [1.8900 ms 1.8961 ms 1.9029 ms]
load_long_package_record time:  [380.69 µs 382.43 µs 384.41 µs]
```